switched to double

+++ Day33_MethodOverLoadingRecap/WarmupTask0.java
@@ -30,14 +30,14 @@ public class WarmupTask0 {
 	}
 
 	//method to remove element at given index of the array
-	public static int[] remove(int[] array, int index){
+	public static double[] remove(double[] array, int index){
 		if(index < 0 || index > array.length - 1){//if the given index number is invalid
 			System.err.println("Invalid index");
 			System.exit(0);
 
 		}
 		//new array with size of original length minus 1
-		int[] result = new int[array.length-1];
+		double[] result = new double[array.length-1];
 		//loop to iterate all indexes
 		for (int i = 0, j = 0; i < array.length; i++) {
 			if (i == index) {//moves all elements of array except the element at given index
